my-website/pi_backend/requirements.txt
flask
requests
flask-cors
python-dotenv
openai
flask_apscheduler
selenium
lxml
bs4